Use KPasswordDialog rather than the pinentry dialog
The pinentry dialog is very minimal and we cannot show custom text on
it. Note that the gpg passphrase callback is run in another thread, so
we cannot show the KPasswordDialog from there. Instead, we first show
the KPasswordDialog (from the main thread), then we start the actual
crypto job whose passphrase callback will just forward the passphrase we
already got.

To pass the passphrase to gpgme we use gpgrt_asprintf, which calls
malloc() under the hood (the buffer will be free()ed later).
That's also what qgpgme does in its tests.