Show error message if passphrase is wrong
A KJob should usually override `errorString()`, but in this case we
actually rely on `errorText()` which is what the plasma applet shows in
the notification body. This way we don't really need to show a
messagebox popup with the error message.

Note that the job could have failed for reasons other than 'wrong
passphrase', but that's unlikely.